[manuf] combine perso 1 & 2 binaries
Since the merging of #21831, larger binaries can now be placed in the
ROM_EXT bootstage. This enables combining the perso binaries into a
single binary, starting by combining binary 1 & 2.
